Add namespacing to functions & constants
All of the current functions are in the top level namespace, which opens
the door for potential naming collisions with other code if / when this
code gets used as a part of a larger project somewhere else. Adding
namespacing makes this substantially more likely.

I've added 2 namespaces, a top level `flashgs` (or `FLASHGS_`) in both
the .h and .cu / .cpp files, as well as an additional anonymous
namespace in the .cu / .cpp files to ensure names of utility functions
don't collide within `flashgs`. The latter is currently unnecessary, but
generally good practice in C++.