Add support for CUDA streams
It's good practice to allow users to specify which streams they'd like
to run CUDA kernels on. I've exposed this functionality through the C++
layer, allowing the stream to be specified at the C++ interfaces. The
default is still stream 0 (the default CUDA stream which was being used
before). Python is also still using stream 0 and does not have the API
exposed to use a different stream.